수정: 커서 기반 페이지네이션 Type 수정 ref: https://github.com/alpha-lens/sb09-hrbank-team01/pull/29#discussion_r2943860254

+++ src/main/java/com/team1/hrbank/dto/request/EmployeeSearchRequest.java
@@ -11,7 +11,7 @@ public record EmployeeSearchRequest(
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate hireDateFrom,
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate hireDateTo,
     String status,
-    Integer idAfter,
+    Long idAfter,
     String cursor,
     Integer size,
     String sortField,
